app/requirements.txt
alembic==1.4.2
amqp==5.0.2
aniso8601==8.0.0
apispec==3.3.0
attrs==19.3.0
Babel==2.8.0
bcrypt==3.1.7
billiard==3.6.3.0
blinker==1.4
cachelib==0.1.1
celery==5.0.5
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.3
coverage==5.1
dateutils==0.6.8
decorator==4.4.2
dnspython==1.16.0
email-validator==1.1.1
Flask==1.1.2
Flask-Babel==1.0.0
Flask-BabelEx==0.9.4
Flask-Cors==3.0.8
Flask-JWT==0.3.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-Session==0.3.2
Flask-SQLAlchemy==2.4.2
Flask-WTF==0.14.3
future==0.18.2
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.1
infinity==1.4
intervals==0.8.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
kombu==5.0.2
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==3.6.0
more-itertools==8.3.0
nose==1.3.7
numpy==1.19.0
packaging==20.4
pandas==1.0.5
passlib==1.7.2
pluggy==0.13.1
prompt-toolkit==3.0.10
psycopg2-binary==2.8.5
py==1.8.1
pycparser==2.20
PyJWT==1.4.2
PyMySQL==1.0.2
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.2
pytest-cov==2.9.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
redis==3.5.3
requests==2.23.0
six==1.15.0
speaklater==1.3
SQLAlchemy==1.3.17
SQLAlchemy-Utils==0.36.6
urllib3==1.25.9
validators==0.15.0
vine==5.0.0
wcwidth==0.1.9
Werkzeug==1.0.1
WTForms==2.3.1
WTForms-Alchemy==0.16.9
WTForms-Components==0.10.4
zipp==3.1.0
